Per #1019, more changes to address SonarQube code smells and reduce the overall number of them lower than what's in the develop branch.

+++ src/basic/vx_util/data_line.cc
@@ -418,7 +418,7 @@ while (it != end) {
 
 if ( Delimiter.find(Line.back()) != string::npos )  Items.emplace_back(""); 
 
-N_items = Items.size();
+N_items = (int) Items.size();
 
 LineNumber = ldf->last_line_number() + 1;
